Finished file restructuring
