+++ components/Footer/Footer.html
@@ -23,7 +23,7 @@
 			</div>
 			<hr class="row-divider-custom"/>
 			<div class="row row-custom">
-					<img src="./assets/img/logo-name.PNG" alt="" class="col-7"/>
+					<img src="/PRNT/assets/img/logo-name.PNG" alt="" class="col-7"/>
 				<div class="col-5 daydate-custom">
 						<p class="text-center daydate-text-custom">
 							Monday–Friday
