+++ pages/contact/script.js
@@ -1,0 +1,2 @@
+import "/PRNT/components/Navbar/Navbar.js";
+import "/PRNT/components/Footer/Footer.js";
